refactor(producer): rename GameDates query folder/namespace to GetGameDates (CR on #506)
Match the sibling convention where each query slice is named after its query
(GetContestOverview, GetContestPlayLog, ...). Renames the folder + namespace for
GetGameDatesQuery / Handler / Validator from ...Queries.GameDates to
...Queries.GetGameDates and updates all references (ContestController, DI). No
behavior change.

(CodeRabbit's other comment — add an IValidator to the API GetLeagueGameDates
proxy handler — skipped: the sibling API proxy handlers use no validator, and
the from<=to invariant is already enforced by the Producer-side validator and
propagated.)
