+++ src/SportsData.Producer/DependencyInjection/ServiceRegistration.cs
@@ -320,8 +320,8 @@ namespace SportsData.Producer.DependencyInjection
             services.AddScoped<FluentValidation.IValidator<GetContestOverviewQuery>, GetContestOverviewQueryValidator>();
             services.AddScoped<FluentValidation.IValidator<GetContestPlayLogQuery>, GetContestPlayLogQueryValidator>();
             services.AddScoped<
-                FluentValidation.IValidator<Application.Contests.Queries.GameDates.GetGameDatesQuery>,
-                Application.Contests.Queries.GameDates.GetGameDatesQueryValidator>();
+                FluentValidation.IValidator<Application.Contests.Queries.GetGameDates.GetGameDatesQuery>,
+                Application.Contests.Queries.GetGameDates.GetGameDatesQueryValidator>();
 
             services.AddScoped<IGroupSeasonsService, GroupSeasonsService>();
             services.AddScoped<ILogoSelectionService, LogoSelectionService>();
@@ -356,8 +356,8 @@ namespace SportsData.Producer.DependencyInjection
             services.AddScoped<IGetMatchupsForSeasonWeekQueryHandler, GetMatchupsForSeasonWeekQueryHandler>();
             services.AddScoped<IGetMatchupByContestIdQueryHandler, GetMatchupByContestIdQueryHandler>();
             services.AddScoped<
-                Application.Contests.Queries.GameDates.IGetGameDatesQueryHandler,
-                Application.Contests.Queries.GameDates.GetGameDatesQueryHandler>();
+                Application.Contests.Queries.GetGameDates.IGetGameDatesQueryHandler,
+                Application.Contests.Queries.GetGameDates.GetGameDatesQueryHandler>();
             services.AddScoped<IGetMatchupsByContestIdsQueryHandler, GetMatchupsByContestIdsQueryHandler>();
             services.AddScoped<IGetMatchupForPreviewQueryHandler, GetMatchupForPreviewQueryHandler>();
             services.AddScoped<IGetMatchupResultQueryHandler, GetMatchupResultQueryHandler>();
